<💄 style>[client:component]:{move header to layout}
