+++ client/app/[locale]/patrol/layout.tsx
@@ -1,0 +1,17 @@
+import Header from "@/components/header";
+
+export default function PatrolLayout({
+    children,
+  }: Readonly<{
+    children: React.ReactNode;
+  }>) {
+    return (
+      <html lang="en">
+        <body
+          className=''>
+            <Header />
+          {children}
+        </body>
+      </html>
+    );
+  }
